Danmaku Nodes
Started work on a more capable NoScript system.

As the old MonoBehaviour implementations of DanmakuSouce, FireModifiers,
and DanmakuControlBehaviours have been removed, it was possible to define
much of the functionality previously with only Danmaku Triggers and
Danmaku Emitters. However the VFW editor, while robust, is not suited for
such involved editing.

Thus the Danmaku Editor has been repurposed as a sole editor for Danmaku
Emitters, which now serve as a container for all of the other components
of the NoScript system.

Previously much of the control was implemented either as a stack (i.e.
Fire Modifiers) or simple lists (Danmaku Controllers). Now the goal is to
get the NoScript system to work as a dataflow graph, starting from Danmaku
Triggers, flowing through consecutive modifier nodes, and into source
nodes that are tied to existing Transforms.

Additions:
- Added IDanmakuObject for all objects depending on DanmakuFields
- Added DanmakuNode, the superclass to all basic nodes within the new
- NoScript system. Each DanmakuNode can have child nodes that it passes
execution flow to after completing it's processing, creating a directed
acyclic graph of data flow instead of single stacks of data.
- Added TriggerNode, a DanmakuNode that starts the whole chain when a
 DanmakuTrigger triggers.
- Added ModifierNode, a DanmakuNode that appends a new Modifier to the chain.
- Added ControllerNode, a DanmakuNode that adds a new DanmakuController to
 the chain.
- Added FireNode, a DanmakuNode that executes the actual fire commands

Changes:
- DanmakuEmitters now implement IEnumerable<DanmakuNode> and can be
- iterated through.
- Moved Danmaku Sources to Modifiers folder since they are now
- FireModifiers.

Removals:
- Removed DanmakuNoScriptContainer, since it is now obsolete (not that it
  was actually functional at any time)