Danmaku Collider Refactors
Danmaku Boundary is now Danmaku Collider, a generic collider with a
filter. Instead of using a HashSet and string filter, it now uses regexes,
which I hope has good performance. Also to save on memory, if the field is
left blank, the regex is left to null instead of allocating an large
collection.

Old functionality that used to be in Danmaku Boundary is now in
DeactivationCollider.

PlayerHitbox is a generic abstraction of existing Player Hitboxes. Using
VFW, it requires that it's parents has an instance of the Danmaku Player
subclass specified by it's generic type constraint.

Other Changes
- Added more overloads to DanmakuField.FindClosest for cleaner code.
- Removed FieldDependentBehaviour since IDanmakuObject exists and is more
  flexible
- The only subclass to FieldDependentBehaviour was Enemy, which now
  subclasses DanmakuCollider, and implements IDanmakuObject
- Enemy initialization is now in Awake, not Start
- Removed DanmakuObjectPrefab and merged it into Danmaku Prefab since
  there doesn't seem to be any need for that abstraction.
- Cleaned up DanmakuPrefab, removed some unneeded stuff