Added more Danmaku Colliders
ConstantForceCollider: pushes danmaku in a certain direction so long as it
is contacting the collider.

AccelerationCollider: speeds up or slows down danmaku so long as it is
contacting the collider.

ReflectionCollider: reflects danmaku using vector reflection.

Changed IDanmakuCollider to pass in a RaycastHit2D parameter as well for
more information on the collision.