Danmaku Field Simplification
Looking for a field with FindClosest will spawn a new one at the origin
without a clipping plane if one currently does not exist.

Also simplified the Field/TargetField mix up. All bullets fired using a
field will be directed toward a TargetField instead. If TargetField is
null, then the field's TargetField is set to itself.