More Optimizations
Danmaku is no longer globally pooled, but instead pooled on a per-prefab
basis, and the pooling is based more on other more successful particle
systems (no queue structures or metadata, just maintains an array index,
and swaps out active and inactive bullets).

Also removed the redundant array accesses in update loops for rendering
and updating danmaku.

Rendering and updating loops are now done in-tandem, to reduce overhead of
iterating over basically the same collection multiple times.

There does seem to be a small bug where otherwise active bullets are being
"paused" and unrenderered when other bullets are destroyed, which should
be fixable.